requirements.txt
fastapi
uvicorn
PyPDF2
langchain>=0.2
langchain-community
langchain_huggingface
sentence-transformers
python-dotenv